TST/REF: implement tests.frame.constructors
